requirements.txt
aiohttp==3.10.11
websockets==12.0
pydantic==2.9.2
PyYAML==6.0.2
python-dotenv==1.0.1
